+++ src/Oro/Bundle/CommerceBundle/DependencyInjection/OroCommerceExtension.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Oro\Bundle\CommerceBundle\DependencyInjection;
 
+use Oro\Bundle\ConfigBundle\DependencyInjection\SettingsBuilder;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -14,6 +15,9 @@ class OroCommerceExtension extends Extension
     #[\Override]
     public function load(array $configs, ContainerBuilder $container): void
     {
+        $config = $this->processConfiguration($this->getConfiguration($configs, $container), $configs);
+        $container->prependExtensionConfig($this->getAlias(), SettingsBuilder::getSettings($config));
+
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yml');
         $loader->load('content_widget_types.yml');